evm config extracted
