+++ ownership-chain/runtime/laos/src/configs/mod.rs
@@ -2,11 +2,11 @@ use crate::{MILLIUNIT, UNIT};
 use ownership_parachain_primitives::Balance;
 
 mod block_rewards_source;
+mod evm;
 pub mod fee;
 mod multisig;
 pub mod parachain_staking;
 mod proxy;
-mod evm;
 
 // Define storage fees as constants for clarity and reuse
 const STORAGE_ITEM_FEE: Balance = 10 * UNIT;
